feat(CampusScheduler): 캠퍼스 운영 시간에 따라 출석 상태 파악 기능 구현

+++ src/main/java/attendance/domain/AttendanceState.java
@@ -18,7 +18,7 @@ public enum AttendanceState {
         this.threshold = threshold;
     }
 
-    public static AttendanceState from(final int diff) {
+    public static AttendanceState from(final long diff) {
         return SORTED_ATTENDANCE_STATE.stream()
                 .filter(attendanceState -> attendanceState.threshold < diff)
                 .findAny()
